+++ packages/acceptance-tests/pkg-tests-specs/sources/features/peerDependenciesMeta.test.ts
@@ -1,3 +1,5 @@
+import {Filename, ppath, xfs} from '@yarnpkg/fslib';
+
 describe(`Features`, () => {
   describe(`peerDependenciesMeta`, () => {
     test(
@@ -41,6 +43,47 @@ describe(`Features`, () => {
           const {stdout} = await run(`install`);
 
           expect(stdout).toMatch(/provides no-deps \(p[0-9a-f]{5}\) with version 1.1.0, which doesn't satisfy what mismatched-peer-deps-lvl0 and some of its descendants request/);
+        },
+      ),
+    );
+
+    test(
+      `it should automatically add corresponding '@types' optional peer dependencies`,
+      makeTemporaryEnv(
+        {
+          dependencies: {
+            foo: `portal:./foo`,
+            "no-deps": `1.0.0`,
+            "@types/no-deps": `1.0.0`,
+          },
+        },
+        async ({path, run, source}) => {
+          await xfs.mkdirPromise(ppath.join(path, `foo` as Filename), {
+            recursive: true,
+          });
+          await xfs.writeJsonPromise(
+            ppath.join(path, `foo` as Filename, Filename.manifest),
+            {
+              name: `foo`,
+              peerDependenciesMeta: {
+                "no-deps": {
+                  optional: true,
+                },
+              },
+            },
+          );
+
+          await expect(run(`install`)).resolves.toMatchObject({
+            code: 0,
+            stdout: expect.not.stringContaining(`YN0002`),
+          });
+
+          await expect(
+            source(`
+                  require
+                    .resolve('@types/no-deps', { paths: [require.resolve('foo/package.json')] })
+                    .replace(/\\\\/g, '/')`),
+          ).resolves.toContain(`/node_modules/@types/no-deps/`);
         },
       ),
     );